+++ MediBalance/Classes/BalanceBoard/BalanceBoard.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using WiimoteLibUni;
+
+namespace MediBalance
+{
+    class BalanceBoard
+    {
+        Wiimote wm = new Wiimote();
+
+    }
+}
