+++ MediBalance/MediBalance.csproj
@@ -97,6 +97,7 @@
     <Compile Include="App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Classes\BalanceBoard\BalanceBoard.cs" />
     <Compile Include="Classes\TestVectors\TestVectors.cs" />
     <Compile Include="MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
@@ -111,6 +112,7 @@
     <None Include="MediBalance_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="lib\BalanceBoard\WiimoteLibUni.dll" />
     <Content Include="lib\MS_Band\Google.Protobuf.dll" />
     <Content Include="lib\MS_Band\Microsoft.Band.dll" />
     <Content Include="lib\MS_Band\Microsoft.Band.Phone_UAP.dll" />
@@ -134,7 +136,9 @@
       <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
+  </ItemGroup>
   <ItemGroup>
     <Reference Include="Google.Protobuf, Version=3.0.0.0, Culture=neutral, PublicKeyToken=a7d26565bac4d604, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
@@ -151,6 +155,10 @@
     <Reference Include="Microsoft.Band.Store_UAP, Version=1.3.20420.1, Culture=neutral, PublicKeyToken=608d7da3159f502b, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>lib\MS_Band\Microsoft.Band.Store_UAP.dll</HintPath>
+    </Reference>
+    <Reference Include="WiimoteLibUni, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>lib\BalanceBoard\WiimoteLibUni.dll</HintPath>
     </Reference>
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
