+++ MediBalance/project.json
@@ -1,6 +1,10 @@
 ﻿{
   "dependencies": {
-    "Microsoft.NETCore.UniversalWindowsPlatform": "5.1.0"
+    "Microsoft.NETCore.UniversalWindowsPlatform": "5.1.0",
+    "Microsoft.OData.Client": "6.13.0",
+    "Microsoft.OData.Core": "6.13.0",
+    "Microsoft.OData.Edm": "6.13.0",
+    "Microsoft.Spatial": "6.13.0"
   },
   "frameworks": {
     "uap10.0": {}
